Components - folder made and productpage- and coctailcorner - files made.
